# openapi/historical-pool-quote-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Historical Pool Quote API",
    "description": "Authenticated historical directed pool quotes from reconstructed Base state.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1"
  },
  "paths": {
    "/coverage": {
      "post": {
        "tags": [
          "coverage"
        ],
        "operationId": "post_coverage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoverageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Continuous reconstruction coverage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is absent or wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Coverage cannot be read safely",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Retry delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/jobs/consistency-check": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "post_consistency_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsistencyCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Existing retained job",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Polling delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmission"
                }
              }
            }
          },
          "202": {
            "description": "New consistency check admitted",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Job polling path"
              },
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Polling delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmission"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is absent or wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Request ID conflicts with retained content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Admission budget exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Waiting queue is full",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Retry delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service cannot safely admit work",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Retry delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/jobs/quote": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "post_quote_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Existing retained job",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Polling delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmission"
                }
              }
            }
          },
          "202": {
            "description": "New quote job admitted",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Job polling path"
              },
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Polling delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmission"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is absent or wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Request ID conflicts with retained content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Comparison budget exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Waiting queue is full",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Retry delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service cannot safely admit work",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Retry delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/jobs/raw-snapshot": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "post_raw_snapshot_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RawSnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Existing retained job",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Polling delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmission"
                }
              }
            }
          },
          "202": {
            "description": "New raw snapshot job admitted",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Job polling path"
              },
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Polling delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmission"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is absent or wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Request ID conflicts with retained content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Waiting queue is full",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Retry delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service cannot safely admit work",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Retry delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/jobs/stored-messages": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "post_stored_messages_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoredMessagesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Existing retained job",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Polling delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmission"
                }
              }
            }
          },
          "202": {
            "description": "New stored messages job admitted",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Job polling path"
              },
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Polling delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmission"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is absent or wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Request ID conflicts with retained content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Waiting queue is full",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Retry delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service cannot safely admit work",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Retry delay in seconds"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Safe error returned by the HTTP API.",
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "details": {},
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Backend": {
        "type": "string",
        "description": "Simulator state partition used by a historical request.",
        "enum": [
          "native",
          "vm",
          "rfq"
        ]
      },
      "BlockBounds": {
        "type": "object",
        "description": "Optional inclusive bounds used by coverage requests.",
        "required": [
          "start",
          "endInclusive"
        ],
        "properties": {
          "endInclusive": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "start": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "CoverageRequest": {
        "type": "object",
        "description": "Request for continuous state reconstruction coverage.",
        "required": [
          "apiRevision",
          "chainId",
          "backends"
        ],
        "properties": {
          "apiRevision": {
            "type": "integer",
            "format": "int32",
            "maximum": 1,
            "minimum": 1
          },
          "backends": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "Simulator state partition used by a historical request.",
              "enum": [
                "native",
                "vm",
                "rfq"
              ]
            },
            "minItems": 1,
            "uniqueItems": true
          },
          "bounds": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BlockBounds"
              }
            ]
          },
          "chainId": {
            "type": "integer",
            "format": "int64",
            "maximum": 8453,
            "minimum": 8453
          }
        },
        "additionalProperties": false
      },
      "CoverageResponse": {
        "type": "object",
        "description": "Continuous reconstructible intervals visible on the read replica.",
        "required": [
          "apiRevision",
          "chainId",
          "backends",
          "visibleThroughBlock",
          "continuousIntervals",
          "knownGaps"
        ],
        "properties": {
          "apiRevision": {
            "type": "integer",
            "format": "int32",
            "maximum": 1,
            "minimum": 1
          },
          "backends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Backend"
            }
          },
          "chainId": {
            "type": "integer",
            "format": "int64",
            "maximum": 8453,
            "minimum": 8453
          },
          "continuousIntervals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockBounds"
            }
          },
          "knownGaps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KnownGap"
            }
          },
          "visibleThroughBlock": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}